Move MTX into own folder
